REF: Loss function position.
