+++ laplax/curv/hessian.py
@@ -1,8 +1,8 @@
 """Hessian vector product for curvature estimation."""
 
 import jax
-import jax.numpy as jnp
 
+from laplax.curv.utils import concatenate_model_and_loss_fn
 from laplax.enums import LossFn
 from laplax.types import (
     Array,
@@ -18,27 +18,6 @@ from laplax.types import (
     TargetArray,
 )
 from laplax.util.tree import mul
-
-
-def log_sigmoid_cross_entropy(
-    logits: Num[Array, "..."], targets: Num[Array, "..."]
-) -> Num[Array, "..."]:
-    """Computes log sigmoid cross entropy given logits and targets.
-
-    This function computes the cross entropy loss between the sigmoid of the logits
-    and the target values. The formula implemented is:
-    -targets * log_sigmoid(logits) - (1 - targets) * log_sigmoid(-logits)
-
-    Args:
-        logits: The predicted logits before sigmoid activation
-        targets: The target values (0 or 1)
-
-    Returns:
-        The computed loss value
-    """
-    return -targets * jax.nn.log_sigmoid(logits) - (1 - targets) * jax.nn.log_sigmoid(
-        -logits
-    )
 
 
 def hvp(func: Callable, primals: PyTree, tangents: PyTree) -> PyTree:
@@ -57,68 +36,6 @@ def hvp(func: Callable, primals: PyTree, tangents: PyTree) -> PyTree:
         The Hessian-vector product.
     """
     return jax.jvp(jax.grad(func), (primals,), (tangents,))[1]
-
-
-def concatenate_model_and_loss_fn(
-    model_fn: ModelFn,  # type: ignore[reportRedeclaration]
-    loss_fn: LossFn | str | Callable,
-    *,
-    has_batch: bool = False,
-) -> Callable[[InputArray, TargetArray, Params], Num[Array, "..."]]:
-    r"""Combine a model function and a loss function into a single callable.
-
-    This creates a new function that evaluates the model and applies the specified
-    loss function. If `has_batch` is `True`, the model function is vectorized over
-    the batch dimension using `jax.vmap`.
-
-    Mathematically, the combined function computes:
-    $L(x, y, \theta) = \text{loss}(f(x, \theta), y)$, where $f$ is the model function,
-    $\theta$ are the model parameters, $x$ is the input, and $y$ is the target.
-
-    Args:
-        model_fn: The model function to evaluate.
-        loss_fn: The loss function to apply. Supported options are:
-            - `LossFn.MSE` for mean squared error.
-            - `LossFn.CROSSENTROPY` for cross-entropy loss.
-            - A custom callable loss function.
-        has_batch: Whether the model function should be vectorized over the batch.
-
-    Returns:
-        A combined function that computes the loss for given inputs, targets, and
-        parameters.
-    """
-    if has_batch:
-        model_fn = jax.vmap(model_fn, in_axes=(0, None))
-
-    if loss_fn == LossFn.MSE:
-
-        def loss_wrapper(
-            input: InputArray, target: TargetArray, params: Params
-        ) -> Num[Array, "..."]:
-            return jnp.sum((model_fn(input, params) - target) ** 2)
-
-        return loss_wrapper
-
-    if loss_fn == LossFn.CROSS_ENTROPY:
-
-        def loss_wrapper(
-            input: InputArray, target: TargetArray, params: Params
-        ) -> Num[Array, "..."]:
-            return log_sigmoid_cross_entropy(model_fn(input, params), target)
-
-        return loss_wrapper
-
-    if callable(loss_fn):
-
-        def loss_wrapper(
-            input: InputArray, target: TargetArray, params: Params
-        ) -> Num[Array, "..."]:
-            return loss_fn(model_fn(input, params), target)
-
-        return loss_wrapper
-
-    msg = f"unknown loss function: {loss_fn}"
-    raise ValueError(msg)
 
 
 def create_hessian_mv_without_data(
